Code fixes to Member vs Participation changes
